rename lib/wrappers/linenoise/linenoise.{c,h}
